+++ src/main/java/net/frozenblock/wilderwild/datagen/tag/WWBlockTagProvider.java
@@ -128,7 +128,9 @@ public final class WWBlockTagProvider extends FabricTagProvider.BlockTagProvider
 			.add(Blocks.DIRT)
 			.add(Blocks.SAND)
 			.add(Blocks.RED_SAND)
+			.add(WWBlocks.SINGED_SAND)
 			.add(WWBlocks.SCORCHED_SAND)
+			.add(WWBlocks.SINGED_RED_SAND)
 			.add(WWBlocks.SCORCHED_RED_SAND);
 	}
 
@@ -312,14 +314,14 @@ public final class WWBlockTagProvider extends FabricTagProvider.BlockTagProvider
 
 		this.getOrCreateTagBuilder(WWBlockTags.SCORCHED_SAND_FEATURE_INNER_REPLACEABLE)
 			.add(Blocks.SAND)
-			.add(WWBlocks.SCORCHED_SAND);
+			.add(WWBlocks.SINGED_SAND);
 
 		this.getOrCreateTagBuilder(WWBlockTags.SCORCHED_SAND_FEATURE_REPLACEABLE)
 			.add(Blocks.SAND);
 
 		this.getOrCreateTagBuilder(WWBlockTags.RED_SCORCHED_SAND_FEATURE_INNER_REPLACEABLE)
 			.add(Blocks.RED_SAND)
-			.add(WWBlocks.SCORCHED_RED_SAND);
+			.add(WWBlocks.SINGED_RED_SAND);
 
 		this.getOrCreateTagBuilder(WWBlockTags.RED_SCORCHED_SAND_FEATURE_REPLACEABLE)
 			.add(Blocks.RED_SAND);
@@ -847,7 +849,9 @@ public final class WWBlockTagProvider extends FabricTagProvider.BlockTagProvider
 			.add(WWBlocks.STONE_CHEST)
 			.add(WWBlocks.NULL_BLOCK)
 			.add(WWBlocks.DISPLAY_LANTERN)
+			.add(WWBlocks.SINGED_SAND)
 			.add(WWBlocks.SCORCHED_SAND)
+			.add(WWBlocks.SINGED_RED_SAND)
 			.add(WWBlocks.SCORCHED_RED_SAND)
 			.add(WWBlocks.GEYSER)
 			.add(WWBlocks.CHISELED_MUD_BRICKS)
@@ -859,7 +863,9 @@ public final class WWBlockTagProvider extends FabricTagProvider.BlockTagProvider
 
 		this.getOrCreateTagBuilder(BlockTags.MINEABLE_WITH_SHOVEL)
 			.addOptionalTag(WWBlockTags.MESOGLEA)
+			.add(WWBlocks.SINGED_SAND)
 			.add(WWBlocks.SCORCHED_SAND)
+			.add(WWBlocks.SINGED_RED_SAND)
 			.add(WWBlocks.SCORCHED_RED_SAND)
 			.add(WWBlocks.TERMITE_MOUND);
 
@@ -890,7 +896,9 @@ public final class WWBlockTagProvider extends FabricTagProvider.BlockTagProvider
 			.add(WWBlocks.RED_MAPLE_LEAVES);
 
 		this.getOrCreateTagBuilder(BlockTags.SAND)
+			.add(WWBlocks.SINGED_SAND)
 			.add(WWBlocks.SCORCHED_SAND)
+			.add(WWBlocks.SINGED_RED_SAND)
 			.add(WWBlocks.SCORCHED_RED_SAND);
 
 		this.getOrCreateTagBuilder(BlockTags.NEEDS_STONE_TOOL)